# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The current server session has been created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "The login fields are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "The credentials are invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Sign in with an existing Roomly account",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "roomly_session": []
          }
        ],
        "summary": "Read the current read-only profile and session expiry",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Roomly API",
    "description": "HTTP API for the Roomly meeting room booking service",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LoginRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "employee@northstar.local"
          },
          "password": {
            "type": "string",
            "example": "EmployeePass!2026",
            "format": "password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Анна Соколова"
          },
          "email": {
            "type": "string",
            "example": "employee@northstar.local"
          },
          "role": {
            "type": "string",
            "enum": [
              "employee",
              "admin"
            ]
          }
        },
        "required": [
          "name",
          "email",
          "role"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserDto"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "user",
          "expiresAt"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "VALIDATION_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Не удалось обработать запрос"
          },
          "fieldErrors": {
            "type": "string",
            "example": [
              "subject: Укажите тему"
            ]
          },
          "requestId": {
            "type": "string",
            "example": "f026e5ca-01e4-4c79-a52e-c44d9c32d967"
          }
        },
        "required": [
          "status",
          "code",
          "message",
          "fieldErrors",
          "requestId"
        ]
      },
...
    }
  }
}